Refactor content_helper panel_helper
